Store boot options needed inside the chroot
This allows these options to be used during the upgrade process as well
as during the build process. Additionally, this allows these boot
options to be specific to a version of the chroot, since they can be
changed when upgrading.

This needs a companion change in the chroot repository to work.