Set live-build mode to progress-linux
The "progress" mode was renamed to "progress-linux" in live-build
3.0~a67.

Since we use an existing chroot, with preinstalled packages and a
configured sources.list, this didn't actually break regular builds.
However, it did break building hdd images, since the default disk label
is based on the mode, but lb_binary_hdd can't handle an empty disk
label.