Partly fix hdd images again
In a recent commit, the call to lb binary_hdd was accidentally removed.
Also, it seems that using ext4 doesn't work in live-build right now (the
allocated image size is too small), so we'll switch back to fat16 for
now.

Note that this does not completely fix disk builds, there seems to be a
problem with the bootloader still.