Restore the syslinux configuration
This links to various parts of the isolinux configuration. Note that,
unlike with isolinux, there is no syslinux.bin for the main syslinux
binary. Instead, this main binary is installed using the "syslinux"
command, which comes from the host.

For this reason, vesamenu.c32 is also linked to the host version and not
the embedded isolinux version, to prevent problems with version
mismatches (at the expense of reintroducing #170 for the hdd version of
the host sylinux version is older than 6).