Add losetup wrapper to force --partscan
Recent versions of Linux, parted or some other bit of software cause
partition devices, like /dev/loop0p1 to be created when running parted
mkpart. However, these devices are not cleaned up when running
losetup -d to remove /dev/loop0 later, so they linger around and confuse
mkfs.

This adds a losetup wrapper that, for invocations where it makes sense
(this is a very rough check), adds the --partscan option, which forces
the kernel to scan for partitions when attaching a loop device. In
practice, this means the lingering partitions are removed when
subsequently attaching the partition itself using the offset option to
losetup.

This is really a workaround until this is fixed in live-build, see
https://bugs.debian.org/865586