Remove host-system patches
There was a script to patch live-build and backport some iso-related
packages from sid (previously these also had patches, that already got
merged into the sid version). Now that stretch is released, all these
patches are included in the stretch versions, so there is no need to
rebuild anything.

This does bump the minimum live-build version. The other packages are
not checked, but with older versions the build should still work,
just not be deterministic.