Work around a problem in syslinux/mcopy
When building a hdd image, syslinux is used. However, there seems to
be some kind of locking or timing issue, that causes the "sylinux"
command to fail:

	P: Copying binary contents into image...
	plain floppy: device "/proc/27467/fd/3" busy (Resource temporarily unavailable):
	Cannot initialize 'S:'
	Bad target s:/ldlinux.sys
	syslinux: failed to create ldlinux.sys

Note that the first two error messages are normal, the "Bad target"
error message is particular to this issue.

I have never figured out what the actual cause is, but it turns out that
adding a second sleep before the mcopy command (which is called by the
syslinux command) prevents this problem from occuring.

This commit adds a wrapper script for the mcopy command that does
exactly this.