Use VBoxManage showhdinfo even when no vdi file exists
This command then shows the UUID from the VirtualBox database, which is
exactly what we need (since that is the UUID that VirtualBox expects to
be able to run without error). If the VDI file is not registered with
VirtualBox, it just shows the UUID from the current file, if any.

This change means the VDI file no longer needs to be kept around, it can
be removed and recreated later without problems.