Update syslinux config for EFI
This allows using the EFI version of syslinux on the hdd images. At the
time of writing, live-build support for syslinux-efi is not merged yet,
but this commit works with custom build of live-build that has
https://salsa.debian.org/live-team/live-build/merge_requests/19 applied
(version of 2019-05-30).

Also note that this does not work (that is, fails at boot with `Undef
symbol FAIL: memset`) with the stretch version of syslinux-efi, but
upgrading the syslinux, syslinux-common and syslinux-efi on the build
host to the buster versions should make things work.

There is no need to explicitly enable syslinux-efi support in the
live-build configuration, since it is enabled by default for hdd images.

When using an unmodified version of live-build (e.g. the Debian/stretch
version), the build still works, but the image will only have BIOS
syslinux support.