Run `lb binary_disk` so `binary/.disk` is created
Without this, binary_iso (from live-build 1:20191221) fails because this
directory is missing. This is due to a change in live-build, for a fix
and a reference to the breaking commit, see
https://salsa.debian.org/live-team/live-build/-/merge_requests/219?

This is a bit of a workaround, and it can be reverted if the above merge
request is merged before bullseye is released (the buster version does
not seem to have this problem).